Move upgradable functionality to a separate source file.
